Add support for storing SetBlocks in Area1
This is used by pool of radiance. The area stored conflicts with
gameOver and picture_fade.

+++ engine/ovr008.cs
@@ -655,7 +655,7 @@ namespace engine
                 {
                     set_value &= 0x7f;
 
-                    ovr031.LoadWalldef(1, set_value & 0xFF);
+                    ovr031.LoadWalldef(1, (short)(set_value & 0xFF));
                 }
             }
             else if (switch_var == 0x324)
@@ -664,7 +664,7 @@ namespace engine
                 {
                     set_value &= 0x7f;
 
-                    ovr031.LoadWalldef(2, set_value & 0xFF);
+                    ovr031.LoadWalldef(2, (short)(set_value & 0xFF));
                 }
             }
             else if (switch_var == 0x326)
@@ -673,7 +673,7 @@ namespace engine
                 {
                     set_value &= 0x7f;
 
-                    ovr031.LoadWalldef(3, set_value & 0xFF);
+                    ovr031.LoadWalldef(3, (short)(set_value & 0xFF));
                 }
             }
         }
